Create attribute_methods.js
Step 2: Methods to add attributes and text content

Now let’s define the attribute methods used in the example. Note the use of the ES2015 rest parameter syntax in addClasses. We’ll also set the element’s textContent property and use shortcut evaluation with the || and && operators for cases where the relevant properties are undefined. That will be relevant when we add methods to create child elements, in case we don’t want to specify all the properties for every child element.